webmin: remove nfsd check from exports-lib.pl
Webmin expects the userspace version of nfsd and attempts to
find its pid when applying a new set of nfs exports. This check
fails because baryon is configured to install the kernelspace
version. In result the command that is assigned to apply_cmd
will never be executed, and the contents of /exports are not
successfully exported.

Fixes [YOCTO #1719]

Signed-off-by: Kevin Strasser <kevin.strasser@linux.intel.com>