orc: specify the generic licenses being used
With the exception of source that contains the Mersenne Twister algorithm,
the majority of orc source code is made available under the BSD 2-Clause
license. The source that contains the Mersenne Twister algorithm is made
available under the BSD 3-Clause license.

This change fixes the following warning:

"WARNING: orc: No generic license file exists for: orc in any provider"

Signed-off-by: Kevin Strasser <kevin.strasser@linux.intel.com>